+++ examples/examples/starwars.rs
@@ -1,3 +1,5 @@
+//! An example of querying the starwars API using the reqwest-blocking feature
+
 mod query_dsl {
     cynic::query_dsl!("examples/starwars.schema.graphql");
 }
@@ -36,17 +38,14 @@ fn main() {
 }
 
 fn run_query() -> cynic::GraphQLResponse<FilmDirectorQuery> {
+    use cynic::http::ReqwestBlockingExt;
+
     let query = build_query();
 
-    let response = reqwest::blocking::Client::new()
+    reqwest::blocking::Client::new()
         .post("https://swapi-graphql.netlify.com/.netlify/functions/index")
-        .json(&query)
-        .send()
-        .unwrap();
-
-    println!("{:?}", response);
-
-    query.decode_response(response.json().unwrap()).unwrap()
+        .run_graphql(query)
+        .unwrap()
 }
 
 fn build_query() -> cynic::Operation<'static, FilmDirectorQuery> {